+++ CHANGELOG.rst
@@ -2,7 +2,6 @@
 
 Changelog
 =========
-
 0.13.1
 ------
 * Model schema now has a discovery API:
@@ -13,6 +12,8 @@ Changelog
   Please see :meth:`tortoise.Tortoise.describe_model` and :meth:`tortoise.Tortoise.describe_models` for more info.
 
 - Fix in generating comments for Foreign Keys in ``MySQL``
+- Added schema support for PostgreSQL. Either set  ``"schema": "custom"`` var in ``credentials`` or as a query parameter ``?schema=custom``
+- Default MySQL charset to ``utf8mb4``. If a charset is provided it will also force the TABLE charset to the same.
 
 0.13.0
 ------